requirements.txt
yfinance>=0.2.36
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.10.0
PyYAML>=6.0
plotly>=5.18.0
streamlit>=1.29.0
pytest>=7.4.0
